Remove or merge all of gateway into reports
