+++ DEVELOPERS.md
@@ -57,11 +57,6 @@ just dev-reset
 
 Access at http://localhost:8000
 
-Login with one of the test user accounts (see Bitwarden entry "Output Explorer NHS Identity Open ID Connect" for password):
-
-- 555036632103
-- 555036633104
-- 555036634105
 
 #### Run tests
