--- gateway/context_processors.py
@@ -1,5 +1,0 @@
-from os import environ
-
-
-def show_login(request):
-    return {"show_login": environ.get("SHOW_LOGIN", False)}
