--- gateway/templates/gateway/login.html
@@ -1,132 +1,0 @@
-{% extends 'base.html' %}
-
-{% block meta %}
-  {% with meta_title="Login | OpenSAFELY: Reports" meta_description="Login" %}
-    <title>{{ meta_title }}</title>
-    <meta name="description" content="{{ meta_description }}">
-    {% include "partials/seo.html" %}
-  {% endwith %}
-{% endblock %}
-
-{% block content %}
-  {% if request.user.is_anonymous %}
-    <div class="bg-gray-50 flex flex-col justify-center py-12 sm:px-6 lg:px-8">
-      <div class="sm:mx-auto sm:w-full sm:max-w-md">
-        <h1 class="mt-6 text-center text-3xl font-extrabold text-gray-900">
-          Login
-        </h1>
-      </div>
-
-      <div class="mt-8 sm:mx-auto sm:w-full sm:max-w-md">
-        <div class="bg-white py-8 px-4 shadow sm:rounded-lg sm:px-10">
-          <form class="space-y-6" method="POST">
-            {% csrf_token %}
-            {% if form.non_field_errors %}
-              <div class="rounded-md bg-red-50 p-4">
-                <div class="flex">
-                  <div class="flex-shrink-0">
-                    {% include "icons/solid/x-circle.svg" with htmlClass="h-5 w-5 text-red-400" %}
-                  </div>
-                  <div class="ml-3">
-                    <h2 class="text-sm font-medium text-red-800">
-                      An error occured
-                    </h2>
-                    <div class="mt-2 text-sm text-red-700">
-                      <ul class="space-y-1">
-                        {% for error in form.non_field_errors %}
-                          <li>{{ error|escape }}</li>
-                        {% endfor %}
-                      </ul>
-                    </div>
-                  </div>
-                </div>
-              </div>
-            {% endif %}
-
-            <div>
-              <label for="id_username" class="block text-sm font-medium text-gray-700">
-                {{ form.username.label }}
-              </label>
-              <div class="mt-1">
-                {{ form.username }}
-              </div>
-              {% if form.username.errors %}
-                {% for error in form.username.errors %}
-                  <p class="mt-2 text-sm font-semibold text-red-700">{{ error|escape }}</p>
-                {% endfor %}
-              {% endif %}
-            </div>
-
-            <div>
-              <label for="id_password" class="block text-sm font-medium text-gray-700">
-                {{ form.password.label }}
-              </label>
-              <div class="mt-1">
-                {{ form.password }}
-              </div>
-              {% if form.username.errors %}
-                {% for error in form.password.errors %}
-                  <p class="mt-2 text-sm font-semibold text-red-700">{{ error|escape }}</p>
-                {% endfor %}
-              {% endif %}
-            </div>
-
-            <div>
-              <button type="submit" class="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-oxford-600 hover:bg-oxford-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-oxford-500">
-                Log in
-              </button>
-            </div>
-          </form>
-
-          {# if the main site login is hidden, this page will be accessed only by internal users, so show username/password options only #}
-          {% if show_login %}
-            <div class="mt-6">
-              <div class="relative">
-                <div class="absolute inset-0 flex items-center">
-                  <div class="w-full border-t border-gray-300"></div>
-                </div>
-                <div class="relative flex justify-center text-sm">
-                  <span class="px-2 bg-white text-gray-500">
-                    Or login with
-                  </span>
-                </div>
-              </div>
-
-              <div class="mt-6">
-                <div>
-                    <a href="{% url "social:begin" "nhsid"  %}" class="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-oxford-600 hover:bg-oxford-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-oxford-500">
-                      NHS Identity Login
-                    </a>
-                </div>
-              </div>
-            </div>
-          {% endif %}
-        </div>
-      </div>
-    </div>
-  {% else %}
-    <div class="container max-w-lg mt-8 mx-auto px-4 md:px-8">
-      <div class="bg-white shadow sm:rounded-lg">
-        <div class="px-4 py-5 sm:p-6">
-          <h1 class="text-lg leading-6 font-medium text-gray-900">
-            Log in
-          </h1>
-          <div class="mt-2 max-w-xl text-sm text-gray-500">
-            <p>
-              You are already logged in as: <strong>
-                {% if request.user.display_name %}
-                  {{ request.user.display_name }}
-                {% else %}
-                  {{ request.user.username }}
-                {% endif %}
-              </strong>
-            </p>
-          </div>
-          <div class="mt-3 text-sm">
-            <a href="/" class="font-medium text-oxford-600 hover:text-oxford-800">Go to the home page <span aria-hidden="true">&rarr;</span></a>
-          </div>
-        </div>
-      </div>
-    </div>
-  {% endif %}
-{% endblock %}
