--- gateway/templates/gateway/organisation_detail.html
@@ -1,6 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-    <h1 class="mt-2">{{ organisation.name }} - {{ organisation.code }}</h1>
-    <p>Placeholder</p>
-{% endblock %}
