--- reports/urls.py
@@ -1,12 +1,0 @@
-from django.urls import path
-from django.views.generic import RedirectView
-
-from .views import report_view
-
-
-app_name = "reports"
-
-urlpatterns = [
-    path("<slug:slug>/", report_view, name="report_view"),
-    path("", RedirectView.as_view(url="/", permanent=True)),
-]
