--- tests/gateway/mocks.py
@@ -1,244 +1,0 @@
-OPENID_CONFIG = {
-    "request_parameter_supported": True,
-    "claims_parameter_supported": False,
-    "introspection_endpoint": "https://dummy-nhs.net/oidc/introspect",
-    "check_session_iframe": "https://dummy-nhs.net/oidc/connect/checkSession",
-    "scopes_supported": [
-        "orgaccess",
-        "universalaccess",
-        "openid",
-        "nationalrbacaccess",
-        "profile",
-        "professionalmemberships",
-        "nhsperson",
-        "odscodes",
-        "organisationalmemberships",
-        "associatedorgs",
-        "email",
-    ],
-    "issuer": "https://dummy-nhs.net/oidc",
-    "id_token_encryption_enc_values_supported": [
-        "A256GCM",
-        "A192GCM",
-        "A128GCM",
-        "A128CBC-HS256",
-        "A192CBC-HS384",
-        "A256CBC-HS512",
-    ],
-    "acr_values_supported": ["AAL1_USERPASS"],
-    "userinfo_encryption_enc_values_supported": [
-        "A256GCM",
-        "A192GCM",
-        "A128GCM",
-        "A128CBC-HS256",
-        "A192CBC-HS384",
-        "A256CBC-HS512",
-    ],
-    "authorization_endpoint": "https://dummy-nhs.net/oidc/authorize",
-    "request_object_encryption_enc_values_supported": [
-        "A256GCM",
-        "A192GCM",
-        "A128GCM",
-        "A128CBC-HS256",
-        "A192CBC-HS384",
-        "A256CBC-HS512",
-    ],
-    "rcs_request_encryption_alg_values_supported": [
-        "RSA-OAEP",
-        "RSA-OAEP-256",
-        "A128KW",
-        "RSA1_5",
-        "A256KW",
-        "dir",
-        "A192KW",
-    ],
-    "claims_supported": [
-        "nhsid_useruid",
-        "mail",
-        "gdc_id",
-        "gmp_id",
-        "consultant_id",
-        "primary_org",
-        "nhsid_org_roles",
-        "gmc_id",
-        "title",
-        "given_name",
-        "uid",
-        "rcn_id",
-        "nhsid_user_orgs",
-        "nhsid_nrbac_roles",
-        "nmc_id",
-        "name",
-        "idassurancelevel",
-        "odscodes",
-        "org_access",
-        "nhsid_uni_roles",
-        "family_name",
-        "gphc_id",
-        "gdp_id",
-    ],
-    "userinfo_signing_alg_values_supported": [
-        "ES384",
-        "HS256",
-        "HS512",
-        "ES256",
-        "RS256",
-        "HS384",
-        "ES512",
-    ],
-    "rcs_request_signing_alg_values_supported": [
-        "ES384",
-        "HS256",
-        "HS512",
-        "ES256",
-        "RS256",
-        "HS384",
-        "ES512",
-    ],
-    "token_endpoint_auth_methods_supported": [
-        "client_secret_post",
-        "private_key_jwt",
-        "self_signed_tls_client_auth",
-        "tls_client_auth",
-        "none",
-        "client_secret_basic",
-    ],
-    "tls_client_certificate_bound_access_tokens": True,
-    "token_endpoint": "https://dummy-nhs.net/oidc/access_token",
-    "response_types_supported": [
-        "code token id_token",
-        "code",
-        "code id_token",
-        "device_code",
-        "id_token",
-        "code token",
-        "none",
-        "token",
-        "token id_token",
-    ],
-    "request_uri_parameter_supported": True,
-    "rcs_response_encryption_enc_values_supported": [
-        "A256GCM",
-        "A192GCM",
-        "A128GCM",
-        "A128CBC-HS256",
-        "A192CBC-HS384",
-        "A256CBC-HS512",
-    ],
-    "userinfo_encryption_alg_values_supported": [
-        "RSA-OAEP",
-        "RSA-OAEP-256",
-        "A128KW",
-        "A256KW",
-        "RSA1_5",
-        "dir",
-        "A192KW",
-    ],
-    "grant_types_supported": ["refresh_token", "authorization_code"],
-    "end_session_endpoint": "https://dummy-nhs.net/oidc/connect/endSession",
-    "rcs_request_encryption_enc_values_supported": [
-        "A256GCM",
-        "A192GCM",
-        "A128GCM",
-        "A128CBC-HS256",
-        "A192CBC-HS384",
-        "A256CBC-HS512",
-    ],
-    "version": "3.0",
-    "rcs_response_encryption_alg_values_supported": [
-        "RSA-OAEP",
-        "RSA-OAEP-256",
-        "A128KW",
-        "A256KW",
-        "RSA1_5",
-        "dir",
-        "A192KW",
-    ],
-    "userinfo_endpoint": "https://dummy-nhs.net/oidc/userinfo",
-    "token_endpoint_auth_signing_alg_values_supported": [
-        "PS384",
-        "ES384",
-        "RS384",
-        "HS256",
-        "HS512",
-        "ES256",
-        "RS256",
-        "HS384",
-        "ES512",
-        "PS256",
-        "PS512",
-        "RS512",
-    ],
-    "require_request_uri_registration": True,
-    "id_token_encryption_alg_values_supported": [
-        "RSA-OAEP",
-        "RSA-OAEP-256",
-        "A128KW",
-        "A256KW",
-        "RSA1_5",
-        "dir",
-        "A192KW",
-    ],
-    "jwks_uri": "https://dummy-nhs.net/oidc/connect/jwk_uri",
-    "subject_types_supported": ["public"],
-    "id_token_signing_alg_values_supported": [
-        "ES384",
-        "HS256",
-        "HS512",
-        "ES256",
-        "RS256",
-        "HS384",
-        "ES512",
-    ],
-    "registration_endpoint": "https://dummy-nhs.net/oidc/register",
-    "request_object_signing_alg_values_supported": [
-        "ES384",
-        "HS256",
-        "HS512",
-        "ES256",
-        "RS256",
-        "HS384",
-        "ES512",
-    ],
-    "request_object_encryption_alg_values_supported": [
-        "RSA-OAEP",
-        "RSA-OAEP-256",
-        "A128KW",
-        "A256KW",
-        "RSA1_5",
-        "dir",
-        "A192KW",
-    ],
-    "rcs_response_signing_alg_values_supported": [
-        "ES384",
-        "HS256",
-        "HS512",
-        "ES256",
-        "RS256",
-        "HS384",
-        "ES512",
-    ],
-}
-
-JWK_KEY = {
-    "kty": "RSA",
-    "d": "ZmswNokEvBcxW_Kvcy8mWUQOQCBdGbnM0xR7nhvGHC-Q24z3XAQWlMWbsmGc_R1o"
-    "_F3zK7DBlc3BokdRaO1KJirNmnHCw5TlnBlJrXiWpFBtVglUg98-4sRRO0VWnGXK"
-    "JPOkBQ6b_DYRO3b0o8CSpWowpiV6HB71cjXTqKPZf-aXU9WjCCAtxVjfIxgQFu5I"
-    "-G1Qah8mZeY8HK_y99L4f0siZcbUoaIcfeWBhxi14ODyuSAHt0sNEkhiIVBZE7QZ"
-    "m-SEP1ryT9VAaljbwHHPmg7NC26vtLZhvaBGbTTJnEH0ZubbN2PMzsfeNyoCIHy4"
-    "4QDSpQDCHfgcGOlHY_t5gQ",
-    "e": "AQAB",
-    "use": "sig",
-    "kid": "testkey",
-    "alg": "RS256",
-    "n": "pUfcJ8WFrVue98Ygzb6KEQXHBzi8HavCu8VENB2As943--bHPcQ-nScXnrRFAUg8"
-    "H5ZltuOcHWvsGw_AQifSLmOCSWJAPkdNb0w0QzY7Re8NrPjCsP58Tytp5LicF0Ao"
-    "Ag28UK3JioY9hXHGvdZsWR1Rp3I-Z3nRBP6HyO18pEgcZ91c9aAzsqu80An9X4DA"
-    "b1lExtZorvcd5yTBzZgr-MUeytVRni2lDNEpa6OFuopHXmg27Hn3oWAaQlbymd4g"
-    "ifc01oahcwl3ze2tMK6gJxa_TdCf1y99Yq6oilmVvZJ8kwWWnbPE-oDmOVPVnEyT"
-    "vYVCvN4rBT1DQ-x0F1mo2Q",
-}
-
-JWK_PUBLIC_KEY = {key: value for key, value in JWK_KEY.items() if key != "d"}
-JWK_PUBLIC_KEYS = [{**JWK_PUBLIC_KEY, "alg": "UNK"}, JWK_PUBLIC_KEY]
